Some more work on CLI and some related changes for speed. 

- suppress redundant info like `installed package foo` on default outputs
- adds `--info` flag to print the extra info without a full verbose (MediumPriority)
- modifies `nimble deps` to print out only direct dependencies and versions
   + this seems more like what people expect when running `nimble deps`
- `nimble deps --tree` prints out the full dedendency tree
- info about url based deps are cached in a `packages_nimblecached.json` PkgList

I did some simple timing on one of my projects and get: 

```sh
# Before cachine URL checks
figuro> nimble deps  2.03s user 0.55s system 26% cpu 9.877 total
# After caching URL checks
figuro> nimble deps  1.38s user 0.24s system 83% cpu 1.931 total
```