@elcritch 2024-12-21
I'd also like to add a `nimble deps --update` and `nimble deps --install` options. Ideally a `nimble deps --update` would be a great way to do network related work but only by explicit request. This is how Elixir's `mix` package manager works and it's great to use. 

@Araq what would you think about adding that?