@Araq 2024-12-21
> @Araq what would you think about adding that?

Fine with me but the naming could be better.