@elcritch 2025-01-09
I ended up removing the URL and pkginfo for URL requires from this since it's a bigger issue and was blocking the UX changes. 

The `nimble deps` is also now colorized, which is nice:

<img width="663" alt="image" src="https://github.com/user-attachments/assets/669c371b-d3f9-4a6b-b5a3-4331b65ebbba" />