@elcritch 2025-01-09
> > @Araq what would you think about adding that?
> 
> Fine with me but the naming could be better.

I left out the `nimble deps --updated` or `nimble deps --install` as well. They'd be nice, whatever others would prefer for names. 

IMHO, in general the CLI options should be hierarchical to aid in discovery and memory. So "deps" options. When you add tab-completion it makes it easy for find things without referring to the docs constantly.