@elcritch 2025-01-09
Also added `nimble list --nimbinaries`

<img width="548" alt="image" src="https://github.com/user-attachments/assets/9563b09a-5f44-4928-86c0-31221ea0910e" />

And colorized the package list and search:

<img width="549" alt="image" src="https://github.com/user-attachments/assets/ff192ee1-2ccf-4fed-9f81-547fe7dc8655" />