@arnetheduck 2025-01-09
Indeed, `deps` is for viewing things - there should at some point exist a command for updating the deps but to have a reasonable UX, it depends on "local-first" mode - otherwise, if you have 2 nimble projects and run "update my deps" in one of them, the _other_ project will also see its deps updated.