@elcritch 2025-01-09
> Indeed, `deps` is for viewing things and we want to keep "view my current state" and "change the state" apart - there should at some point exist a command for updating the deps but to have a reasonable UX, it depends on "local-first" mode - otherwise, if you have 2 nimble projects and run "update my deps" in one of them, the _other_ project will also see its deps updated.

That's a good point. I found Elixir's mix layout to make sense and to still keep view and modify distinct. They use a `mix deps.<command> --opts` format. In Nimble that'd be like having `nimble depsUpdate` and a `nimble deps`. It's a nice balance of discoverability and separation of concerns. Mix is one of the least problematic PMs I've used, which is why I reference it.

> it depends on "local-first" mode - otherwise, if you have 2 nimble projects and run "update my deps" in one of them, the _other_ project will also see its deps updated

Hmmm, looks like we do have `nimble update` which seems equivalent to Mix's `mix deps.update xyz`. However whenever you install deps with Mix they're installed locally and creates a lock file. So it's a different model.